--- src/main/java/br/com/hideyoshi/auth/base/auth/model/RoleToUserDTO.java
@@ -1,10 +1,0 @@
-package br.com.hideyoshi.auth.base.auth.model;
-
-import lombok.Data;
-
-@Data
-public
-class RoleToUserDTO {
-    private String username;
-    private String roleName;
-}
