--- src/main/java/br/com/hideyoshi/auth/base/auth/repo/UserRepository.java
@@ -1,14 +1,0 @@
-package br.com.hideyoshi.auth.base.auth.repo;
-
-import br.com.hideyoshi.auth.base.auth.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-
-    Optional<User> findByEmail(String email);
-}
