+++ src/main/java/br/com/hideyoshi/auth/base/interceptor/ConfigInterceptor.java
@@ -1,0 +1,18 @@
+package br.com.hideyoshi.auth.base.interceptor;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Component
+@RequiredArgsConstructor
+public class ConfigInterceptor implements WebMvcConfigurer {
+
+    private final UserResourceAccessInterceptor userResourceAccessInterceptor;
+
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(userResourceAccessInterceptor);
+    }
+
+}
