+++ src/main/java/br/com/hideyoshi/auth/base/model/RoleToUserDTO.java
@@ -1,0 +1,10 @@
+package br.com.hideyoshi.auth.base.model;
+
+import lombok.Data;
+
+@Data
+public
+class RoleToUserDTO {
+    private String username;
+    private String roleName;
+}
