+++ src/main/java/br/com/hideyoshi/auth/base/repository/UserRepository.java
@@ -1,0 +1,14 @@
+package br.com.hideyoshi.auth.base.repository;
+
+import br.com.hideyoshi.auth.base.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+
+    Optional<User> findByEmail(String email);
+}
