--- src/main/java/br/com/hideyoshi/auth/healthchecker/api/HealthCheckerController.java
@@ -1,23 +1,0 @@
-package br.com.hideyoshi.auth.healthchecker.api;
-
-
-import br.com.hideyoshi.auth.util.guard.UserResourceGuard;
-import br.com.hideyoshi.auth.util.guard.UserResourceGuardEnum;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-
-@Log4j2
-@Controller
-public class HealthCheckerController {
-    @RequestMapping("/health")
-    @UserResourceGuard(accessType = UserResourceGuardEnum.OPEN)
-    public ResponseEntity<String> healthCheck() {
-        log.info("Health check requested");
-        return ResponseEntity.ok("Health check successful!");
-    }
-}
